requirements.txt
psycopg2-binary==2.9.9
SQLAlchemy==1.4.42
pandas==1.3.5
alembic==1.10.3
apache-airflow==2.3.4
google-cloud-bigquery == 3.13.0
pyarrow== 12.0.1



